Makefile: fix phony targets
clean, run and debug were marked as PHONY targets in the Makefile, but
unfortunately due to a misspelling (as PHONEY) they didn't work as
proper phony targets. If a file named 'clean' was created, for example,
the clean target ceased to work:

> touch clean
> make clean
make: 'clean' is up to date.

Correct the misspelling to fix the phony targets.

Signed-off-by: Andreas Bofjall <andreas@gazonk.org>